Archived Tickets
Archived tickets have been properly implemented, with the new DateTime variable in-place for all tickets. [This is done inside the FileHandling file]

There is also now a CANCELLED status if a staff member decides that they don't want to continue with their ticket. [A part of the TicketView file]

Tickets are now sorted before they are displayed, the most recent ones go first.