ArchivedDT Change
I've added a new Ticket field called ArchivedDT, which is the time that a ticket is set to archived,

I've also run small modifications of the code for better performance.

I've also, also, created a WriteSuccess() method to let the user know of successful stuff, rather than only having an Error writer.